chore: replace old ReportForm
